+++ src/main/java/insane96mcp/insanesurvivaloverhaul/mixin/accessor/BlockBehaviourAccessor.java
@@ -2,10 +2,12 @@ package insane96mcp.insanesurvivaloverhaul.mixin.accessor;
 
 import net.minecraft.world.level.block.state.BlockBehaviour;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Mutable;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
 @Mixin(BlockBehaviour.class)
 public interface BlockBehaviourAccessor {
     @Accessor
+    @Mutable
     void setIsRandomlyTicking(boolean isRandomlyTicking);
 }