+++ src/main/java/insane96mcp/insanesurvivaloverhaul/mixin/module/misc/potionandeffects/PoisonMobEffectMixin_PotionsAndEffects.java
@@ -1,0 +1,14 @@
+package insane96mcp.insanesurvivaloverhaul.mixin.module.misc.potionandeffects;
+
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
+import insane96mcp.insanesurvivaloverhaul.module.misc.PotionsAndEffects;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(targets = "net.minecraft.world.effect.PoisonMobEffect")
+public class PoisonMobEffectMixin_PotionsAndEffects {
+    @ModifyExpressionValue(method = "shouldApplyEffectTickThisTick", at = @At(value = "CONSTANT", args = "intValue=25"))
+    public int onPoisonTickDamage(int poisonFactor) {
+        return PotionsAndEffects.getPoisonDamageSpeed(poisonFactor);
+    }
+}